research/manuscript_Reader/with_kiro/requirements.txt
# Telugu OCR System - Complete Requirements
# This file contains all dependencies for full functionality

# Core dependencies
pillow>=10.0.0
opencv-python-headless>=4.8.0
pdf2image>=3.1.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0.1
scipy>=1.10.0
scikit-image>=0.20.0

# OCR engines - Local
pytesseract>=0.3.13
easyocr>=1.7.2

# OCR engines - Advanced (optional, large downloads)
paddleocr>=2.7.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0

# Cloud OCR APIs (optional)
google-cloud-vision>=3.4.0
azure-cognitiveservices-vision-computervision>=0.9.0
boto3>=1.28.0
google-auth>=2.20.0

# Text processing and language support
python-bidi>=0.4.2
shapely>=2.0.0
pyclipper>=1.3.0
nltk>=3.8.0
python-Levenshtein>=0.20.0
fuzzywuzzy>=0.18.0

# Image preprocessing
scikit-image>=0.20.0
imageio>=2.30.0

# Output formats
openpyxl>=3.1.0
python-docx>=0.8.11
jinja2>=3.1.0
markdown>=3.4.0

# Visualization and reporting
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and utilities
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0
colorama>=0.4.6
pathlib2>=2.3.7
configparser>=5.3.0

# System monitoring
psutil>=5.9.0
memory-profiler>=0.60.0

# Development tools (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation (optional)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# API and web utilities
requests>=2.31.0
urllib3>=2.0.0
certifi>=2023.7.22

# File handling
glob2>=0.7
pathspec>=0.11.0

# Logging and debugging
rich>=13.0.0

# Note: Some packages may require system-level dependencies:
# 
# For Ubuntu/Debian:
# sudo apt-get update
# sudo apt-get install tesseract-ocr tesseract-ocr-tel
# sudo apt-get install poppler-utils
# sudo apt-get install libgl1-mesa-glx  # For OpenCV
# 
# For macOS:
# brew install tesseract tesseract-lang
# brew install poppler
# 
# For Windows:
# Download Tesseract from: https://github.com/UB-Mannheim/tesseract/wiki
# Download poppler from: https://blog.alivate.com.au/poppler-windows/
# 
# GPU Support (optional, for faster processing):
# For CUDA 11.8: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121